Add code for mnist temporal

+++ nonlinear/source/mnist_hqrc.py
@@ -34,16 +34,18 @@ MNIST_DIR = "../mnist"
 RES_MNIST_DIR = "../results/rs_mnist"
 MNIST_SIZE="10x10"
 
-def dump_reservoir_states(tempfile, Xs, qparams, model, init_rs, ranseed):
-    rstates = []
-    for s in Xs:
-        input_signals = np.array(s)
-        output_signals = model.init_forward(qparams, input_signals, init_rs=init_rs, ranseed = ranseed)
-        rstates.append(output_signals.ravel())
-    rstates = np.array(rstates)
-    with open(tempfile, 'wb') as wrs:
-        pickle.dump(rstates, wrs)
-    return tempfile
+def dump_reservoir_states(logger, datfile, Xs, qparams, model, init_rs, ranseed):
+    if os.path.isfile(datfile) == True:
+        logger.debug('File existed {}'.format(datfile))
+    else:
+        input_signals = np.array(Xs)
+        rstates = model.init_forward(qparams, input_signals, init_rs=init_rs, ranseed = ranseed)
+        #_, rstates =  model.feed_forward(input_signals, predict=False, use_lastrho=use_lastrho)
+        rstates = np.array(rstates)
+        with open(datfile, 'wb') as wrs:
+            pickle.dump(rstates, wrs)
+            logger.debug('Dumped file {}'.format(datfile))
+    return datfile
 
 if __name__  == '__main__':
     # Check for command line arguments
@@ -62,9 +64,10 @@ if __name__  == '__main__':
     parser.add_argument('--nqrs', type=int, default=1, help='Number of reservoirs')
     parser.add_argument('--ntrials', type=int, default=1)
     parser.add_argument('--virtuals', type=int, default=1)
-    parser.add_argument('--strengths', type=float, default=0.0, help='Connection strengths')
-    parser.add_argument('--tau', type=float, default=4.0, help='Input interval')
-
+    parser.add_argument('--strengths', type=str, default='0.0', help='Connection strengths')
+    parser.add_argument('--taudeltas', type=str, default='-7,-6,-5,-4,-3,-2,-1,0,1,2,3,4,5,6,7')
+    parser.add_argument('--interval', type=float, default=0.2, help='tau-interval')
+    
     parser.add_argument('--linear_reg', type=int, default=0)
     parser.add_argument('--use_corr', type=int, default=0)
     parser.add_argument('--full', type=int, default=0)
@@ -80,12 +83,17 @@ if __name__  == '__main__':
     print(args)
 
     n_qrs, n_spins, beta, rseed = args.nqrs, args.spins, args.beta, args.rseed
-    J, init_rho, tau, V, alpha = args.coupling, args.rho, args.tau, args.virtuals, args.strengths
+    J, init_rho, V = args.coupling, args.rho, args.virtuals
     
     solver, linear_reg, use_corr, transient = args.solver, args.linear_reg, args.use_corr, args.transient
     full_mnist, label1, label2 = args.full, args.label1, args.label2
     ntrials, dynamic, savedir = args.ntrials, args.dynamic, args.savedir
     mnist_dir, mnist_size, nproc = args.mnist_dir, args.mnist_size, args.nproc
+
+    taudeltas = [float(x) for x in args.taudeltas.split(',')]
+    #taudeltas = list(np.arange(-7, 7.1, args.interval))
+    taudeltas = [2**x for x in taudeltas]
+    strengths = [float(x) for x in args.strengths.split(',')]
 
     if os.path.isfile(savedir) == False and os.path.isdir(savedir) == False:
         os.mkdir(savedir)
@@ -108,159 +116,135 @@ if __name__  == '__main__':
     if os.path.isdir(bindir) == False:
         os.mkdir(bindir)
 
-    basename = '{}_{}_linear_{}_nqrs_{}_corr_{}_nspins_{}_V_{}_tau_{}_alpha_{}'.format(\
-        mnist_size, dynamic, linear_reg, n_qrs, use_corr, n_spins, V, tau, alpha)
-    log_filename = os.path.join(logdir, '{}.log'.format(basename))
-    logger = get_module_logger(__name__, log_filename)
-    logger.info(log_filename)
-
-    x_train, y_train_lb, x_test, y_test_lb, x_val, y_val_lb = gen_mnist_dataset(mnist_dir, mnist_size)
+    basename = 'join_{}_{}_linear_{}_nqrs_{}_corr_{}_nspins_{}_V_{}'.format(\
+        mnist_size, dynamic, linear_reg, n_qrs, use_corr, n_spins, V)
     
+    x_train, y_train_lb, x_test, y_test_lb = gen_mnist_dataset_join_test(mnist_dir, mnist_size)
+    imlength = int(x_train.shape[1] / n_qrs)
+
     if full_mnist > 0:
-        Y_train = np.identity(10)[y_train_lb]
-        Y_test  = np.identity(10)[y_test_lb]
-        Y_val   = np.identity(10)[y_val_lb]
+        Y_train_org = np.identity(10)[y_train_lb]
+        Y_test_org  = np.identity(10)[y_test_lb]
     else:
+        basename = '{}_lb_{}_{}'.format(basename, label1, label2)
         # Train data
         train_ids = (y_train_lb == label1) | (y_train_lb == label2)
-        # x_train = x_train[train_ids, :]
+        x_train = x_train[train_ids, :]
         y_train_lb = y_train_lb[train_ids]
         y_train_lb[y_train_lb == label1] = 0
         y_train_lb[y_train_lb == label2] = 1
 
-        Y_train = np.identity(2)[y_train_lb]
+        Y_train_org = np.identity(2)[y_train_lb]
 
         # Test data 
         test_ids = (y_test_lb == label1) | (y_test_lb == label2)
-        # x_test = x_test[test_ids, :]
+        x_test = x_test[test_ids, :]
         y_test_lb = y_test_lb[test_ids]
         y_test_lb[y_test_lb == label1] = 0
         y_test_lb[y_test_lb == label2] = 1
 
-        Y_test  = np.identity(2)[y_test_lb]
+        Y_test_org  = np.identity(2)[y_test_lb]
 
-        # Val data
-        val_ids = (y_val_lb == label1) | (y_val_lb == label2)
-        # x_val = x_val[val_ids, :]
-        y_val_lb = y_val_lb[val_ids]
-        y_val_lb[y_val_lb == label1] = 0
-        y_val_lb[y_val_lb == label2] = 1
+    log_filename = os.path.join(logdir, '{}.log'.format(basename))
+    logger = get_module_logger(__name__, log_filename)
+    logger.info(log_filename)
+    logger.info('shape x_train={}, y_train={},  x_test={}, y_test={}'.format(\
+        x_train.shape, Y_train_org.shape, x_test.shape, Y_test_org.shape))
 
-        Y_val  = np.identity(2)[y_val_lb]
+    jobs, pipels = [], []
+    datfs = dict()
+                
+    for alpha in strengths:
+        for tau in taudeltas:
+            # Create params and model
+            qparams = QRCParams(n_units=n_spins-1, n_envs=1, max_energy=J,\
+                beta=beta, virtual_nodes=V, tau=tau, init_rho=init_rho, solver=solver, dynamic=dynamic)
+            model = hqrc.HQRC(nqrc=n_qrs, gamma=alpha, sparsity=1.0, sigma_input=1.0, use_corr=use_corr)
+            model.init_reservoir(qparams, ranseed=rseed)
+            init_rs, rs_seed, use_lastrho = False, 0, False
+            # create training and testing data
+            if linear_reg <= 0:
+                for datlb in ['train', 'test']:
+                    if datlb == 'train':
+                        Xt = x_train
+                    else:
+                        Xt  = x_test
+                    # Make Xs to put in hqrc
+                    Xs = [x.reshape(n_qrs, -1) for x in Xt]
+                    Xs = np.concatenate(Xs, axis=1)
+                    # Create training/test file
+                    (ns, ds) = Xt.shape
+                    logger.info('{} file in time series {}'.format(datlb, Xs.shape))
+                    datfile = '{}_{}_alpha_{:.2f}_tau_{:.4f}_shape_{}_{}.bin'.format(datlb, basename, alpha, tau, ns, ds)
+                    datfile = os.path.join(bindir, datfile)
+                    datfs['{}_alpha_{:.2f}_tau_{:.4f}'.format(datlb, alpha, tau)] = datfile
+                    p = multiprocessing.Process(target=dump_reservoir_states, \
+                        args=(logger, datfile, Xs, qparams, model, init_rs, rs_seed))
+                    jobs.append(p)
+    # Start the process
+    for p in jobs:
+        p.start()
 
+    # Ensure all processes have finished execution
+    for p in jobs:
+        p.join()
+
+    # Sleep 5s
+    time.sleep(5)
     
-    logger.info('shape x_train={}, y_train={},  x_test={}, y_test={}, x_val={}, y_val={}'.format(\
-        x_train.shape, Y_train.shape, x_test.shape, Y_test.shape, x_val.shape, Y_val.shape))
-
-    # Create params and model
-    qparams = QRCParams(n_units=n_spins-1, n_envs=1, max_energy=J,\
-        beta=beta, virtual_nodes=V, tau=tau, init_rho=init_rho, solver=solver, dynamic=dynamic)
-    model = hqrc.HQRC(nqrc=n_qrs, gamma=alpha, sparsity=1.0, sigma_input=1.0, use_corr=use_corr)
-    model.init_reservoir(qparams, ranseed=rseed)
-    init_rs, rs_seed = False, 0
-    # create training and testing data
-    if linear_reg <= 0:
-        datfs = dict()
-        for datlb in ['train', 'test', 'val']:
-            if datlb == 'train':
-                Xt = x_train
-            else:
-                Xt  = x_test
-            # Make Xs to put in hqrc
-            Xs = [x.reshape(n_qrs, -1) for x in Xt]
-            # Create training/test file
-            (ns, ds) = Xt.shape
-            datfile = '{}_{}_{}_{}.bin'.format(datlb, basename, ns, ds)
-            datfile = os.path.join(bindir, datfile)
-            datfs[datlb] = datfile
-            if os.path.isfile(datfile) == True:
-                logger.debug('File existed {}'.format(datfile))
-                continue
-
-            lst = np.array_split(Xs, nproc)
-            print(len(lst), lst[0].shape)
-            jobs, pipels = [], []
-            for proc_id in range(nproc):
-                Xsmall = lst[proc_id]
-                if Xsmall.size == 0:
-                    continue
-                tempfile = os.path.join(bindir, '{}_{}.bin'.format(datlb, proc_id))
-                p = multiprocessing.Process(target=dump_reservoir_states, \
-                    args=(tempfile, Xsmall, qparams, model, init_rs, rs_seed))
-                jobs.append(p)
-                pipels.append(tempfile)
-            
-            # Start the process
-            for p in jobs:
-                p.start()
-
-            # Ensure all processes have finished execution
-            for p in jobs:
-                p.join()
-
-            # Sleep 5s
-            time.sleep(5)
-
-            # join the training file
-            X_states = []
-            for filename in pipels:
-                if os.path.isfile(filename) == True:
-                    with open(filename, 'rb') as rrs:
-                        X_states.append(pickle.load(rrs))
-                    # Delete file
-                    os.remove(filename)
-            X_states = np.concatenate(X_states)
-            logger.info('Concatenate {} file with shape {}'.format(datlb, X_states.shape))
-            # dump to datfile
-            with open(datfile, 'wb') as wrs:
-                pickle.dump(X_states, wrs)
-            
-
     # perform regression
-    if linear_reg > 0:
-        X_train = np.array(x_train)
-        if full_mnist == 0:
-            X_train = X_train[train_ids, :]
+    for alpha in strengths:
+        for tau in taudeltas:
+            if linear_reg > 0:
+                X_train = np.array(x_train)
+                if full_mnist == 0:
+                    X_train = X_train[train_ids, :]
+            else:
+                # Training
+                with open(datfs['train_alpha_{:.2f}_tau_{:.4f}'.format(alpha, tau)], 'rb') as rrs:
+                    X_train = pickle.load(rrs)
+                    # if use_corr == 0:
+                    #     ids = np.array(range(X_train.shape[1]))
+                    #     ids = ids[ids % 15 < 5]
+                    #     X_train = X_train[:, ids]
+                    #if full_mnist == 0:
+                    #    X_train = X_train[train_ids, :]
+            Y_train = np.repeat(Y_train_org, imlength, axis=0)
+            X_train = np.hstack( [X_train, np.ones([X_train.shape[0], 1]) ] )
+            logger.info('Nqr={}, V={}, alpha={}, tau={}, X_train shape={}, Y_train shape={}'.format(\
+                n_qrs, V, alpha, tau, X_train.shape, Y_train.shape))
 
-    else:
-        # Training
-        with open(datfs['train'], 'rb') as rrs:
-            X_train = pickle.load(rrs)
-            # if use_corr == 0:
-            #     ids = np.array(range(X_train.shape[1]))
-            #     ids = ids[ids % 15 < 5]
-            #     X_train = X_train[:, ids]
-            if full_mnist == 0:
-                X_train = X_train[train_ids, :]
+            #XTX = X_train.T @ X_train
+            #XTY = X_train.T @ Y_train
+            #I = np.identity(np.shape(XTX)[1])	
+            #pinv_ = scipypinv2(XTX + beta * I)
+            #W_out = pinv_ @ XTY
+            W_out = np.linalg.pinv(X_train, rcond = beta) @ Y_train
+            logger.info('Wout shape={}'.format(W_out.shape))
+            y_train_predict = group_avg(X_train @ W_out, imlength)
+            logger.info('y_train_predict shape={}'.format(y_train_predict.shape))
+            
+            train_acc = get_acc(y_train_predict, y_train_lb)
+            logger.info('Nqr={}, V={}, tau={}, alpha={}, Train acc={}'.format(n_qrs, V, tau, alpha, train_acc))
 
-    X_train = np.hstack( [X_train, np.ones([X_train.shape[0], 1]) ] )
-    logger.info('V={}, tauB={}, X_train shape={}'.format(V, tau, X_train.shape))
-
-    XTX = X_train.T @ X_train
-    XTY = X_train.T @ Y_train
-    I = np.identity(np.shape(XTX)[1])	
-    pinv_ = scipypinv2(XTX + beta * I)
-    W_out = pinv_ @ XTY
-    logger.info('Wout shape={}'.format(W_out.shape))
-    train_acc = get_acc(X_train @ W_out, y_train_lb)
-    logger.info('Train acc={}'.format(train_acc))
-
-    # Testing
-    if linear_reg > 0:
-        X_test = np.array(x_test)
-        if full_mnist == 0:
-            X_test = X_test[test_ids, :]
-    else:
-        with open(datfs['test'], 'rb') as rrs:
-            X_test = pickle.load(rrs)
-            # if use_corr == 0:
-            #     ids = np.array(range(X_test.shape[1]))
-            #     ids = ids[ids % 15 < 5]
-            #     X_test = X_test[:, ids]
-            if full_mnist == 0:
-                X_test = X_test[test_ids, :]
-
-    X_test = np.hstack( [X_test, np.ones([X_test.shape[0], 1]) ] )
-    logger.info('V={}, tau={}, alpha={}, X_test shape = {}'.format(V, tau, alpha, X_test.shape))
-    test_acc = get_acc(X_test @ W_out, y_test_lb)
-    logger.info('Test acc={}'.format(test_acc))
+            # Testing
+            if linear_reg > 0:
+                X_test = np.array(x_test)
+                if full_mnist == 0:
+                    X_test = X_test[test_ids, :]
+            else:
+                with open(datfs['test_alpha_{:.2f}_tau_{:.4f}'.format(alpha, tau)], 'rb') as rrs:
+                    X_test = pickle.load(rrs)
+                    # if use_corr == 0:
+                    #     ids = np.array(range(X_test.shape[1]))
+                    #     ids = ids[ids % 15 < 5]
+                    #     X_test = X_test[:, ids]
+                    #if full_mnist == 0:
+                    #    X_test = X_test[test_ids, :]
+            Y_test = np.repeat(Y_test_org, imlength, axis=0)
+            X_test = np.hstack( [X_test, np.ones([X_test.shape[0], 1]) ] )
+            logger.info('Nqr={}, V={}, tau={}, alpha={}, X_test shape = {}'.format(n_qrs, V, tau, alpha, X_test.shape))
+            y_test_predict = group_avg(X_test @ W_out, imlength)
+            logger.info('y_test_predict shape={}'.format(y_test_predict.shape))
+            test_acc = get_acc(y_test_predict, y_test_lb)
+            logger.info('Nqr={}, V={}, tau={}, alpha={}, Test acc={}'.format(n_qrs, V, tau, alpha, test_acc))
